% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.zz_endpoint}
\alias{.zz_endpoint}
\title{An auxiliary object that holds endpoints for various functions}
\usage{
.zz_endpoint()
}
\description{
An auxiliary object that holds endpoints for various functions
}
\keyword{internal}
