\name{fkoh111-package}
\alias{fkoh111-package}
\alias{fkoh111}
\docType{package}
\title{
\packageTitle{fkoh111}
}
\description{
\packageDescription{fkoh111}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fkoh111}
\packageIndices{fkoh111}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fkoh111}

Maintainer: \packageMaintainer{fkoh111}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
