% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_config.R
\name{zz_config}
\alias{zz_config}
\title{Handler for Zamzar configs.}
\usage{
zz_config()
}
\value{
A list containing attributes from config.json.
}
\description{
Per default zz_configs expects an inst folder in the current working directory (initialized via getwd()).
and looks for a config.json with the appropriate attributes.
}
\examples{
zz_config()
}
