% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_delete.R
\name{zz_delete}
\alias{zz_delete}
\title{Delete file from Zamzar endpoint}
\usage{
zz_delete(id = NULL, usr = NULL)
}
\arguments{
\item{id}{The target id for a file you wish to delete.}

\item{usr}{The username/API key you are using. If not set, the function
will check if a key exists as a `ZAMZAR_USR` variable  in `.Renviron` and use that.    

See: \url{https://developers.zamzar.com/user}}
}
\value{
A status message indicating either success or failure.
}
\description{
Delete file from Zamzar endpoint via id.
Per default `zz_delete()` assumes that you're doing development, thus using a 
development endpoint. Set prod boolean to `TRUE` to change this behavior.
}
\details{
Please note that a Zamzar key passed as argument to `usr` takes precedence over a
Zamzar key extracted from an `.Renviron`.
}
\examples{
\dontrun{
# An example of zz_delete() with a hardcoded file id
# Please note that zz_delete doesn't need to dinstinguish betweend endpoints (prod or dev)

zz_get(id = 12345678)

}
}
