% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{zz_endpoint}
\alias{zz_endpoint}
\title{An auxiliary object to hold endpoints that are being used by various functions}
\usage{
zz_endpoint()
}
\value{
Nuttin
}
\description{
An auxiliary object to hold endpoints that are being used by various functions
}
\keyword{internal}
