% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_format.R
\name{zz_format}
\alias{zz_format}
\title{Simple wrapper for Zamzar endpoint returning a list of formats you can
convert between}
\usage{
zz_format(usr = NULL)
}
\arguments{
\item{usr}{The username/API key you are using for Zamzar.
See: https://developers.zamzar.com/user}
}
\value{
A list of formats
}
\description{
Simple wrapper for Zamzar endpoint returning a list of formats you can
convert between
}
\examples{
zz_format()
}
