% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_get.R
\name{zz_get}
\alias{zz_get}
\title{Get file from Zamzar endpoint}
\usage{
zz_get(id = NULL, usr = NULL, name = NULL, extension = NULL, prod = FALSE)
}
\arguments{
\item{id}{The target id for a previously passed file. Most likely
returned from zz_get_info().}

\item{usr}{The username/API key you are using for Zamzar.
See: \url{https://developers.zamzar.com/user}}

\item{name}{The name of the file you are fetching from Zamzar. If a name is
not assigned to the file, then we're using the id as name.}

\item{extension}{The extension of the file you are fetching from Zamzar. 
If an extension is not assigned, then we're using .png extension.}

\item{prod}{Boolean deciding whether to use a production endpoint or
a development endpoint. Defaults to FALSE (That is, development endpoint).}
}
\value{
A file written to disk
}
\description{
Get file from Zamzar endpoint
}
\examples{
\donttest{
zz_get()
}
}
