% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_get_info.R
\name{zz_get_info}
\alias{zz_get_info}
\title{Get Zamzar info}
\usage{
zz_get_info(usr = NULL, latest = TRUE)
}
\arguments{
\item{usr}{The username/API key you are using for Zamzar.
See: \url{https://developers.zamzar.com/user}}

\item{latest}{Boolean deciding whether or not we should only return the 
latest target id. If switched to false, will return a list of all assigned
target ids}
}
\value{
A target id or alternatively a list of target ids
}
\description{
Get Zamzar info
}
\examples{
\donttest{
zz_get_info()
}
}
