% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_post.R
\name{zz_post}
\alias{zz_post}
\title{Post file to Zamzar endpoint}
\usage{
zz_post(file = NULL, extension = NULL, usr = NULL, prod = FALSE)
}
\arguments{
\item{file}{The path to the file you want to convert.}

\item{extension}{The file type you want to convert to. E.g., "png".}

\item{usr}{The username/API key you are using for Zamzar.
See: \url{https://developers.zamzar.com/user}}

\item{prod}{Boolean deciding whether to use prod or dev endpoint.
Defaults to FALSE (That is, dev endpoint).}
}
\value{
A response object as defined by httr::message_for_status()
}
\description{
Post file to Zamzar endpoint
}
\examples{
\donttest{

# Per default zz_post uses the development endpoint
zz_post(file = "avatar.emf", usr = "passwd", extension = "png")

# Setting prod param to FALSE is the same as above
zz_post(file = "avatar.emf", usr = "passwd", extension = "png", prod = FALSE)

# You need to flip prod to TRUE if you want to use the production endpoint
zz_post(file = "avatar.emf", usr = "passwd", extension = "png", prod = TRUE)

}
}
\seealso{
[zz_format()] for a list of formats you can convert to.
}
