% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from.R
\name{import_from}
\alias{import_from}
\title{Import script files for the R programming language from a given folder.}
\usage{
import_from(src, strict = TRUE)
}
\arguments{
\item{src}{The the folder containg .R files that you want to import.
This can either be the name of a folder, or the relative path to a given folder.
If no folder or path is provided, the current wd is imported.}

\item{strict}{Evaluate .R extensions in strict mode, that is, source only
script files with a capital .R extension. Defaults to TRUE.}
}
\value{
Returns object(s) assigned in .R files located in the folder imported
by import_from
}
\description{
Per default import_from will only import files with a capitalized
extension, that is an .R extension, and not files with a lowercase
.r exension.
}
\details{
This is in accordance with general guidelines in the community, as well as
advised by authorities within the community, e.g.:
Wickham, H. (2015). Advanced R . Boca Raton, FL: CRC Press.
}
\examples{
import_from("R")

}
