% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_folder.R
\name{init_folder}
\alias{init_folder}
\title{Init a a given folder containing .R files}
\usage{
init_folder(src)
}
\arguments{
\item{src}{which is a path or a folder name}
}
\value{
nuttin expect for more objects in your environment
}
\description{
This function initializes files with a proper .R extension.
}
\examples{
init_folder("R")
}
