% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_format.R
\name{zz_format}
\alias{zz_format}
\title{Get vector of accepted formats from Zamzar}
\usage{
zz_format(usr = NULL, origin = NULL)
}
\arguments{
\item{usr}{The username/API key you are using for Zamzar. If not set, zz_format()
will see if a key exists as ZAMZAR_USR variable  in .Renviron and use that.

See: \url{https://developers.zamzar.com/user}}

\item{origin}{The origin format you want to convert from.
If a valid param is passed to origin, zz_format() returns a list of \strong{targets} and \strong{costs}.
\itemize{
\item \code{targets}: A vector containing the formats your origin can be converted to.
\item \code{costs}: A vector containing the cost for converting between your origin and a given target.
}

See also: \url{https://developers.zamzar.com/formats}

If no orign param is passed to zz_format(), a character vector containing all
the accepted formats for the \strong{origin} param is returned.

See: \url{https://developers.zamzar.com/formats}}
}
\value{
Either a list of formats from the API that you can convert to,
or a character vector of accepted origin formats.
}
\description{
Get vector of accepted formats from Zamzar
}
\section{Details}{

Please note that a Zamzar key passed as usr param takes precedence over a
Zamzar key extracted from the .Renviron.
}

\examples{
\donttest{
# Returns a character vector of all the accepted formats for the origin param
zz_format(usr = "key")

# Same as above (assuming a valid key in .Renviron)
zz_format()

# Returns an error since the origin param isn't recognized by the Zamzar API
zz_format(origin = "invalid_origin")

# Returns a list of targets that origin can be converted to, and of the cost of
# converting to a given target.
zz_format(origin = "emf")
}
}
