% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_format.R
\name{zz_format}
\alias{zz_format}
\title{Get formats from Zamzar}
\usage{
zz_format(usr = NULL, origin = NULL)
}
\arguments{
\item{usr}{The username/API key you are using for Zamzar.

See: \url{https://developers.zamzar.com/user}}

\item{origin}{The origin format you want to convert.
If param origin is passed, zz_format() returns a list of $targets and of $costs.
The target vector contains tagets that you can convert your origin to.
The cost vector contains the cost for conversion between your origin and target.  

See: \url{https://developers.zamzar.com/formats}

If no orign param has been passed to zz_format(), a character type of
all the origin formats accepted by the Zamzar API is returned.  

See: \url{https://developers.zamzar.com/formats}}
}
\value{
Either a list of formats from the API that you can convert to,
or a type character of all the origin formats accepted by the Zamzar API
}
\description{
Get formats from Zamzar
}
\examples{
\donttest{
zz_format(usr = "key")

zz_format(usr = "key", origin = "invalid_origin")

zz_format(usr = "key", origin = "emf")
}
}
