% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_get.R
\name{zz_get}
\alias{zz_get}
\title{Get file from Zamzar endpoint}
\usage{
zz_get(id = NULL, usr = NULL, name = NULL, extension = NULL, prod = FALSE)
}
\arguments{
\item{id}{The target id for a previously passed file. Most likely
returned from zz_get_info().}

\item{usr}{The username/API key you are using for Zamzar. If not set, zz_get()
will see if a key exists as `ZAMZAR_USR` variable  in .Renviron and use that.    

See: \url{https://developers.zamzar.com/user}}

\item{name}{The name of the file you are fetching from Zamzar. If a name is
not assigned to the file, then we're using the id as file name.}

\item{extension}{The extension of the file you are fetching from Zamzar.}

\item{prod}{Boolean deciding whether to use a production endpoint or
a development endpoint. Defaults to FALSE (That is, development endpoint).}
}
\value{
A file written to disk
}
\description{
Get file from Zamzar endpoint
}
\section{Details}{

Get file from Zamzar endpoint via id.
Per default zz_get() assumes that you're doing development, thus using a 
development endpoint. Set prod bool to TRUE to change this behaviour.

Please note that a Zamzar key passed as usr param takes precedence over a
Zamzar key extracted from the .Renviron.
}

\examples{
\donttest{
# An example of zz_get() utilized with hardcoded arguments
zz_get(id = 12345678, usr = "key", name = "my_avatar", extension = "png")

# An example zz_get() used in conjunction with zz_get_info()
response <- zz_get_info(usr = "key", latest = TRUE)
zz_get(usr = usr, id = response$id, extension = response$extension, prod = TRUE)
}
}
