% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_get.R
\name{zz_get}
\alias{zz_get}
\title{Simple wrapper for Zamzar endpoint to fetch a file from a target id}
\usage{
zz_get(target_id = NULL, usr = NULL, prod = FALSE)
}
\arguments{
\item{target_id}{The target id for a previously passed file. Most likely
returned from zz_get_id()}

\item{usr}{The username/API key you are using for Zamzar.
See: https://developers.zamzar.com/user}

\item{prod}{Boolean deciding whether to use a production endpoint or
a development endpoint. Defaults to FALSE (That is, development endpoint).}
}
\value{
A file written to disk
}
\description{
Per default zz_get() assumes that you're doing development, thus using a 
development endpoint. Set prod bool to TRUE to change this behaviour.
}
\examples{
zz_get()
}
