% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_get_id.R
\name{zz_get_id}
\alias{zz_get_id}
\title{Simple wrapper for Zamzar endpoint accepting images for conversion}
\usage{
zz_get_id(usr = NULL, prod = FALSE, latest = TRUE)
}
\arguments{
\item{usr}{The username/API key you are using for Zamzar. See: https://developers.zamzar.com/user}

\item{prod}{Boolean deciding whether to use a production endpoint or
a development endpoint. Defaults to FALSE (That is, development endpoint).}

\item{latest}{Boolean deciding whether or not we should only return the 
latest id. If switched to false, will return a list of all assigned ids}
}
\value{
An id or alternatively a list of ids
}
\description{
Per default zz_post assumes that you're doing development, thus using a 
development endpoint. Set prod bool to TRUE to change this behavious.
}
\examples{
zz_get_id()
}
