% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_get_info.R
\name{zz_get_info}
\alias{zz_get_info}
\title{Get info from Zamzar}
\usage{
zz_get_info(usr = NULL, latest = TRUE)
}
\arguments{
\item{usr}{The username/API key you are using for Zamzar.

See: \url{https://developers.zamzar.com/user}}

\item{latest}{Boolean deciding whether or not zz_get_info() should solely
return attributes for the latest target id.

If switched to FALSE, zz_get_info() will return attributes for all files
that have been sumitted to the API within a reasonable timeframe.

The returned list contains the following attributes:
\itemize{
\item \code{id}: The unique file identifier assigned to a file by Zamzar.
\item \code{extension}: The extension representing the format of the file.
\item \code{created_at}: The time at which the file was created at the Zamzar servers.
}}
}
\value{
A list of file attributes.
}
\description{
Get info from Zamzar
}
\section{zz_get_info}{

Get info on files submitted to the Zamzar API via your current token

Per default zz_get_info() assumes you want information for the last
submitted file. To get information on all the files that have been submitted
within a reasonable timeframe, set bool latest to FALSE.

Please note: objects returned from zz_get_info() doesn't differentiate
between files that have been assigned to either the development or
production endpoint. You have to keep track of this yourself.
}

\examples{
\donttest{
# Provided a valid token, will return a list of attributes
zz_get_info(usr = "key")

# Provided a valid token, will return a list of files
# submitted to the API within a reasonable timeframe.
zz_get_info(usr = "key", latest = FALSE)

}
}
