% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_get_info.R
\name{zz_get_info}
\alias{zz_get_info}
\title{Simple wrapper for Zamzar endpoint returning an id(s) for a posted file(s)}
\usage{
zz_get_info(usr = NULL, latest = TRUE)
}
\arguments{
\item{usr}{The username/API key you are using for Zamzar.
See: \url{https://developers.zamzar.com/user}}

\item{latest}{Boolean deciding whether or not we should only return the 
latest target id. If switched to false, will return a list of all assigned
target ids}
}
\value{
A target id or alternatively a list of target ids
}
\description{
Per default zz_get_info() assumes that you only want the target id for the latest 
assigned file. Set latest bool to FALSE to change this behaviour.
}
\details{
Please note: zz_get_info() doesn't differentiate between files that have been
assigned to either the development or production endpoint.
Thus you have to keep track of this yourself.
}
\examples{
\donttest{
zz_get_info()
}
}
