% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_post.R
\name{zz_post}
\alias{zz_post}
\title{Simple wrapper for Zamzar endpoint accepting images for conversion}
\usage{
zz_post(file = NULL, target = NULL, usr = NULL, prod = FALSE)
}
\arguments{
\item{file}{The file you want to convert. Potentially the path to the file
you want to convert.}

\item{target}{The file you want to convert to. E.g., "png".}

\item{usr}{The username/API key you are using for Zamzar.
See: https://developers.zamzar.com/user}

\item{prod}{Boolean deciding whether to use a production endpoint or
a development endpoint. Defaults to FALSE (That is, development endpoint).}
}
\value{
A response object
}
\description{
Per default zz_post() assumes that you're doing development, thus using a 
development endpoint. Set prod bool to TRUE to change this behaviour.
}
\examples{
zz_post()
}
